+++ core/src/jsMain/kotlin/zakadabar/stack/frontend/builtin/layout/ZkLayoutTheme.kt
@@ -10,6 +10,7 @@ class ZkLayoutTheme(
     val paddingStep: Int = 10,
     val sliderColor: String = ZkColors.LightBlue.a700,
     val sliderHoverColor: String = ZkColors.LightBlue.a400,
+    val defaultForeground: String = ZkColors.black,
     val defaultBackground: String = ZkColors.Gray.c100,
     val titleBarHeight: Int = 44
 )