+++ core/src/jsMain/kotlin/zakadabar/stack/frontend/builtin/theme/ZkBuiltinDarkTheme.kt
@@ -14,15 +14,23 @@ import zakadabar.stack.frontend.builtin.modal.ZkModalTheme
 import zakadabar.stack.frontend.builtin.sidebar.ZkSideBarTheme
 import zakadabar.stack.frontend.builtin.table.ZkTableTheme
 import zakadabar.stack.frontend.builtin.toast.ZkToastTheme
+import zakadabar.stack.frontend.resources.ZkColors
 import zakadabar.stack.frontend.resources.ZkTheme
 
 open class ZkBuiltinDarkTheme : ZkTheme {
-    override var button = ZkButtonTheme()
+    override var button = ZkButtonTheme(
+        background = ZkColors.Design.navBlue
+    )
     override var dock = ZkDockTheme()
-    override var font = ZkFontTheme()
+    override var font = ZkFontTheme(
+        family = "'IBM Plex Sans', sans-serif"
+    )
     override var form = ZkFormTheme()
     override var icon = ZkIconTheme()
-    override var layout = ZkLayoutTheme()
+    override var layout = ZkLayoutTheme(
+        defaultForeground = ZkColors.Design.gray2,
+        defaultBackground = ZkColors.Design.gray9
+    )
     override var modal = ZkModalTheme()
     override var sidebar = ZkSideBarTheme()
     override var tabContainer = ZkTabContainerTheme()