+++ lib/markdown/src/jsMain/kotlin/zakadabar/lib/frontend/markdown/MarkdownStyles.kt
@@ -7,7 +7,7 @@ import zakadabar.stack.frontend.application.ZkApplication
 import zakadabar.stack.frontend.resources.ZkColors
 import zakadabar.stack.frontend.resources.css.ZkCssStyleSheet
 
-class MarkdownStyles : ZkCssStyleSheet<MarkdownStyles>(ZkApplication.theme) {
+class MarkdownStyles : ZkCssStyleSheet() {
 
     val codeFence by cssClass {
         borderRadius = 4